pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apify-client>=1.10.0",
    "fastapi>=0.115.12",
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.2",
    "notion-client>=2.3.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "structlog>=25.3.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.34.2",
]
